+++ src/view/ViewCallbacks.hh
@@ -6,8 +6,8 @@
 namespace pmines {
     namespace view {
         struct ViewCallbacks {
-            std::function<void(size_t, size_t)> tile_left_clicked;
-            std::function<void(size_t, size_t)> tile_right_clicked;
+            std::function<void(int, int)> tile_left_clicked;
+            std::function<void(int, int)> tile_right_clicked;
         };
     }
 }
